+++ src/BattleSession.hpp
@@ -1,8 +1,6 @@
 #pragma once
 class PlayLayer;
 class GJBaseGameLayer;
-class PlayerObject;
-class GameObject;
 namespace versus::battle {
 void begin(PlayLayer* layer);
 bool activeFor(GJBaseGameLayer* layer);
@@ -11,7 +9,6 @@ bool blocksInput(GJBaseGameLayer* layer);
 bool blocksPause(GJBaseGameLayer* layer);
 void sample(PlayLayer* layer);
 void died(PlayLayer* layer);
-void survivedLethalHit(PlayLayer* layer, PlayerObject* player, GameObject* hazard);
 void completed(PlayLayer* layer);
 bool beforeReset(PlayLayer* layer);
 void afterReset(PlayLayer* layer);