+++ src/MatchLaunch.cpp
@@ -515,7 +515,6 @@ class $modify(VersusMatchPlayLayer, PlayLayer) {
         PlayLayer::destroyPlayer(player, object);
         if (effective && (effective == m_player1 || (m_gameState.m_isDualMode && effective == m_player2))) {
             if (effective->m_isDead) versus::battle::died(this);
-            else versus::battle::survivedLethalHit(this, effective, object);
         }
     }
     void levelComplete() {